+++ R/names.R
@@ -4,9 +4,17 @@ model_names <- c("mutau" = "Aggregate data (with control group)",
                  "rubin" = "Rubin model with aggregate data",
                  "rubin_full" = "Rubin model with individual-level data",
                  "quantiles" = "Effects on quantiles",
-                 "logit" = "Logistic model (individual-level data)",
+                 "logit" = "Logistic model",
                  "sslab" = "Spike & slab model (3 components)")
 
+data_type_names <- c("pool_noctrl_narrow" = "Aggregate (effects only)",
+                     "pool_wide" = "Aggregate (control and effects)",
+                     "pool_binary" = "Aggregate (event counts)",
+                     "individual" = "Individual-level with continuous outcome",
+                     "individual_binary" = "Individual-level with binary outcome")
+
+
+# Default settings for matching models with input data types (convert_inputs)
 model_data_types <- c("rubin" = "pool_noctrl_narrow",
                       "mutau" = "pool_wide",
                       "logit" = "individual_binary",
@@ -16,11 +24,14 @@ model_data_types <- c("rubin" = "pool_noctrl_narrow",
                       #for now no quantiles model from summary level data
                       "quantiles" = "individual")
 
-data_type_names <- c("pool_noctrl_narrow" = "Aggregate (effects only)",
-                     "pool_wide" = "Aggregate (control and effects)",
-                     "individual" = "Individual-level with continuous outcome",
-                     "individual_binary" = "Individual-level with binary outcome")
+data_type_default_model <- c("pool_noctrl_narrow" = "rubin",
+                             "pool_wide" = "mutau",
+                             "pool_binary" = "logit",
+                             "individual" = "rubin_full",
+                             "individual_binary" = "logit")
 
+
+# Allowed priors for different types of variables
 available_priors <- list(
   "real" = c("normal", "uniform", "cauchy"),
   "positive_real" = c("normal", "uniform", "cauchy"),